scalability-tools/simulator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fistsim"
version = "0.1.0"
dependencies = [
    "docker == 6.1.3",
    "PyYAML == 6.0.1",
    "Requests == 2.31.0",

]